requirements.txt
fastapi
uvicorn
geopandas
shapely
lxml
pyproj
pandas
google-cloud-functions
functions-framework
mangum
httpx
python-dotenv